Setup button removed. Process automatized (unity 2018)
APPC-738

The setup is being runned everytime the developer imports the plugin or changes a variable on the prefab. Build & Run works now correctly without the android custom build window or the setup button (unity 2018 only)
